// apps/fusion/project/ws/app/src/lib/routes/profile/routes/analytics/models/analytics.model.ts
export namespace NSAnalyticsData {
// ...
  export interface ITopContent {
    lex_id: string
    count: number
    content_name: string
    progress_range: IProgressRange[]
  }
  export interface ILearningHistory {
    lex_id: string
    content_name: string
    progress: number
    last_progress_date: string
  }
  export interface ILearningHistoryProgress {
    [key: string]: IProgressRange[]
  }

  export interface IGoalsShared {
    goal_id: string
    goal_title: string
    email_id: string
    last_updated_on: string
    progress: number
    goal_type: string
    goal_end_date: null
    goal_start_date: null
    shared_on: string
    created_on: string
    shared_by: string
    status: string
  }
  export interface IPlayListProgress {
    last_updated_on: string
    email_id: string
    play_list_id: string
    progress: number
    visibility: string
    shared_by: string
    play_list_title: string
    created_on: string
  }
  export interface IPlayListShared {
    type: string
    shared_by: string
    play_list_title: string
    email_id: string
    play_list_id: string
    resource_id: string
    created_on: string
    progress: number
  }
  export interface IUserProgressResponse {
    top_content_jl: ITopContent[]
    top_content_unit: ITopContent[]
    learning_history: ILearningHistory[]
    learning_history_progress_range: ILearningHistoryProgress
    goal_progress: never[]
    goal_shared_by_me: IGoalsShared[]
    goal_shared_to_me: IGoalsShared[]
    playlist_progress: IPlayListProgress[]
    playlist_shared_by_me: IPlayListShared[]
    playlist_shared_to_me: IPlayListShared[]
  }

}